ci: reuse the result of the previous step in the same E2E hardware flow
